tls: make pub available tls configurations and utilities
